+++ sphinx/source/changelog.rst
@@ -22,6 +22,9 @@ Changelog
   ``suzy-soft2`` tape-sampling loop accelerators
 * Added the ``Screen`` configuration parameter for :ref:`trace.py <trace-conf>`
   (to specify whether to display screen contents)
+* Added the :ref:`rstHandler` component
+* Added the ``--handle-rst`` option to :ref:`sna2ctl.py` (for activating the
+  :ref:`rstHandler`)
 * Fixed how :ref:`tap2sna.py` handles ``--tape-stop`` for PZX files that
   contain STOP blocks